cloud-run/demo/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
google-cloud-bigquery==3.0.1
pytz==2022.1
